extract class around channel_source_map table
