+++ linen-modern/src/main/scala/x7c1/linen/modern/accessor/SettingSourceAccessor.scala
@@ -2,8 +2,8 @@ package x7c1.linen.modern.accessor
 
 import android.database.sqlite.SQLiteDatabase
 import android.database.{Cursor, SQLException}
-import x7c1.linen.modern.accessor.database.SourceParts
-import x7c1.linen.modern.accessor.unread.{ChannelSourceMapParts, SourceRatingParts}
+import x7c1.linen.modern.accessor.database.{ChannelSourceMapParts, SourceParts}
+import x7c1.linen.modern.accessor.unread.SourceRatingParts
 import x7c1.linen.modern.struct.Date
 import x7c1.wheat.macros.database.{TypedFields, TypedCursor}
